requirements.txt
# FormatFusion - Universal File Converter
# Python dependencies

# GUI Framework
PyQt6>=6.4.0

# Image processing
Pillow>=9.0.0

# Video/Audio processing
ffmpeg-python>=0.2.0

# Additional utilities
pathlib2>=2.3.0; python_version < "3.4"
